requirements.txt
certifi==2020.4.5.1
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
opt-einsum==3.2.1
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.3.1
python-dateutil==2.8.1
six==1.14.0
torch==1.5.0
tqdm==4.46.0
